drivers: display: mb_display: Make display context more compact
Reorder the struct members for more compact layout. The current row
tracker tracks just values 0 to 2 so uint8_t is more than enough for
it.


Former-commit-id: 86e64b77f22c9b6c98d9fd6f2a2f372070cf94b3

+++ drivers/display/mb_display.c
@@ -7,7 +7,7 @@ DECL|SCROLL_OFF|macro|SCROLL_OFF
 DECL|SCROLL_START|macro|SCROLL_START
 DECL|clear_display|function|static void clear_display(struct k_timer *timer)
 DECL|col_mask|variable|col_mask
-DECL|cur|member|int cur; /* Currently rendered row */
+DECL|cur|member|uint8_t cur; /* Currently rendered row */
 DECL|dev|member|struct device *dev; /* GPIO device */
 DECL|display|variable|display
 DECL|duration|member|int32_t duration; /* Duration for each shown image */
